requirements.txt
opencv-python>=4.8
numpy>=1.24
pandas>=2.0
matplotlib>=3.7
transformers>=4.40
torch>=2.0
torchvision>=0.15
Pillow>=10.0
ultralytics>=8.0
